fix tests to work with jpa, change id type in model classes to long

+++ src/main/java/pl/futurecollars/invoicing/db/memory/InMemoryDatabase.java
@@ -10,19 +10,19 @@ import pl.futurecollars.invoicing.model.Invoice;
 public class InMemoryDatabase implements Database {
 
     private final HashMap<Integer, Invoice> invoices = new HashMap<>();
-    private int index = 1;
+    private long index = 1;
 
     @Override
     public long save(Invoice invoice) {
         invoice.setId(index);
-        invoices.put(index, invoice);
+        invoices.put((int) index, invoice);
         return index++;
     }
 
     @Override
     public Optional<Invoice> getById(long id) {
 
-        return Optional.ofNullable(invoices.get(id));
+        return Optional.ofNullable(invoices.get((int) id));
     }
 
     @Override
@@ -40,7 +40,7 @@ public class InMemoryDatabase implements Database {
 
     @Override
     public Optional<Invoice> delete(long id) {
-        return Optional.ofNullable(invoices.remove(id));
+        return Optional.ofNullable(invoices.remove((int) id));
 
     }
